Handle MSVC warnings

+++ include/librealsense2/hpp/rs_export.hpp
@@ -87,8 +87,8 @@ namespace rs2
             std::vector<rs2::vertex> new_verts;
             std::vector<vec3d> normals;
             std::vector<std::array<uint8_t, 3>> new_tex;
-            std::map<int, int> idx_map;
-            std::map<int, std::vector<vec3d>> index_to_normals;
+            std::map<size_t, size_t> idx_map;
+            std::map<size_t, std::vector<vec3d>> index_to_normals;
 
             new_verts.reserve(p.size());
             if (use_texcoords) new_tex.reserve(p.size());
@@ -112,11 +112,11 @@ namespace rs2
             auto profile = p.get_profile().as<video_stream_profile>();
             auto width = profile.width(), height = profile.height();
             static const auto threshold = get_option(OPTION_PLY_THRESHOLD);
-            std::vector<std::array<int, 3>> faces;
+            std::vector<std::array<size_t, 3>> faces;
             if (mesh)
             {
-                for (int x = 0; x < width - 1; ++x) {
-                    for (int y = 0; y < height - 1; ++y) {
+                for (size_t x = 0; x < width - 1; ++x) {
+                    for (size_t y = 0; y < height - 1; ++y) {
                         auto a = y * width + x, b = y * width + x + 1, c = (y + 1)*width + x, d = (y + 1)*width + x + 1;
                         if (verts[a].z && verts[b].z && verts[c].z && verts[d].z
                             && fabs(verts[a].z - verts[b].z) < threshold && fabs(verts[a].z - verts[c].z) < threshold
